requirements.txt
certifi==2022.9.24
elasticsearch==7.11.0
numpy==1.23.5
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
urllib3==1.26.12
